+++ Changelog.md
@@ -2,6 +2,7 @@
 
 - Updated dependencies, thanks to Alex Touchet for his help!
 - Increased minimum Rust version to 1.67 as this is required by the latest versions of our dependencies
+- Fixed panic in complex to real vector operations, https://github.com/liebharc/basic_dsp/issues/51
 
 
 ## Version 0.10.0